requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
django-searchable-encrypted-fields==0.2.1
djangorestframework==3.14.0
idna==3.4
install==1.3.5
pip-autoremove==0.10.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.18.0
Pympler==1.0.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
